+++ operator/internal/resources/aistore/cmn/job.go
@@ -21,7 +21,7 @@ const CleanupPrefix = "cleanup"
 func NewCleanupJob(ais *aisv1.AIStore, nodeName string) *batchv1.Job {
 	ttl := int32(0) // delete the pod as soon as it is completed
 	jobName := fmt.Sprintf("%s-%s-", CleanupPrefix, strings.ReplaceAll(nodeName, ".", "-"))
-	hostpathPrefix := *ais.Spec.StateStorageHostPathPrefix()
+	stateDir := StateHostPath(ais, "")
 	return &batchv1.Job{
 		ObjectMeta: metav1.ObjectMeta{
 			GenerateName: jobName,
@@ -32,8 +32,8 @@ func NewCleanupJob(ais *aisv1.AIStore, nodeName string) *batchv1.Job {
 			Template: corev1.PodTemplateSpec{
 				Spec: corev1.PodSpec{
 					Affinity:      createNodeAffinitySpec(nodeName),
-					Containers:    createContainerSpec(hostpathPrefix),
-					Volumes:       createVolumeSpec(hostpathPrefix),
+					Containers:    createContainerSpec(stateDir),
+					Volumes:       createVolumeSpec(stateDir),
 					Tolerations:   ais.GetAllTolerations(),
 					RestartPolicy: corev1.RestartPolicyNever,
 				},
@@ -64,16 +64,16 @@ func createNodeAffinitySpec(nodeName string) *corev1.Affinity {
 }
 
 // createContainerSpec constructs the container spec for the job
-func createContainerSpec(hostpathPrefix string) []corev1.Container {
+func createContainerSpec(stateDir string) []corev1.Container {
 	return []corev1.Container{
 		{
 			Name:    "cleanup",
 			Image:   "docker.io/aistorage/ais-operator-helper:latest",
-			Command: []string{"/cleanup-helper", "-dir=" + hostpathPrefix},
+			Command: []string{"/cleanup-helper", "-dir=" + stateDir},
 			VolumeMounts: []corev1.VolumeMount{
 				{
 					Name:      "hostpath",
-					MountPath: hostpathPrefix,
+					MountPath: stateDir,
 				},
 			},
 		},
@@ -81,13 +81,13 @@ func createContainerSpec(hostpathPrefix string) []corev1.Container {
 }
 
 // createVolumeSpec constructs the volume spec for the job
-func createVolumeSpec(hostpathPrefix string) []corev1.Volume {
+func createVolumeSpec(stateDir string) []corev1.Volume {
 	return []corev1.Volume{
 		{
 			Name: "hostpath",
 			VolumeSource: corev1.VolumeSource{
 				HostPath: &corev1.HostPathVolumeSource{
-					Path: hostpathPrefix,
+					Path: stateDir,
 				},
 			},
 		},